+++ Runtime/Services/LocalizationService/Runtime/ILocalizationChangeNotifier.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Services.LocalizationService.Runtime
+{
+    public interface ILocalizationChangeNotifier
+    {
+        public event Action Changed;
+        void Initialize();
+        void Dispose();
+    }
+}